refactor routes to pull out content and add withFileName
